Various int -> size_t fixes

+++ src/command.h
@@ -15,9 +15,9 @@ typedef struct {
 
 // parse-command.c
 char *parse_command_arg(const char *cmd, bool tilde);
-int find_end(const char *cmd, int pos, Error **err);
+ssize_t find_end(const char *cmd, size_t pos, Error **err);
 bool parse_commands(PointerArray *array, const char *cmd, Error **err);
-char **copy_string_array(char **src, int count);
+char **copy_string_array(char **src, size_t count);
 
 // run.c
 extern const Command *current_command;
